requirements.txt
Flask
Flask-Cors
mysql-connector-python
gunicorn
jsonify
datetime
pytz
werkzeug
boto3
bcrypt
pyjwt
python-dotenv
google-api-python-client